FWBUG-160: cujo-luaoop-debugging: Fix serialization class
The issue here is if tables are reused, i.e. address they point to is
the same, lookup in history table returns a sting and it gets written
into a buffer, instead of a call to a corresponding writer.

Now writer is called for a given luatype regardless of label used.